+++ apps/blog/src/app/components/pages/home/index.tsx
@@ -1,9 +1,5 @@
 import { Blogs } from "./components/Blogs";
 
 export function HomePage() {
-  return (
-    <>
-      <Blogs />
-    </>
-  );
+  return <Blogs />;
 }
